requirements.txt
Django==1.6.2
django-tinymce==1.5.2
wsgiref==0.1.2
